Караван пылит

+++ SatchelApp/src/app/shared/components/routers/cart-page/cart-page.component.ts
@@ -3,7 +3,7 @@ import { CartPageService } from 'src/app/core/services/cart-page.service';
 import { UserService } from 'src/app/core/services/user.service';
 import { Product } from '../catalog/catalog.component';
 import { Router} from '@angular/router';
-import { ConfigService } from 'src/app/core/services/config.service';
+import { ConfigService, PaymentType, ShippingType } from 'src/app/core/services/config.service';
 
 @Component({
   selector: 'app-cart-page',
@@ -20,11 +20,17 @@ export class CartPageComponent {
   shoppingCart : Product[] = [];
   finalPrice: number = 0
   paymentTypeId: number = 0; 
+  shippingTypeId: number = 0;
+  paymentTypes: PaymentType[] = [];
+  shippingTypes: ShippingType[] = [];
+  statusMsg = '';
 
   ngOnInit() {
     this.checkAuthorization();
     //ngOnInit будет прерван, если chechAuth не пройдет (интересный факт)
     this.getShoppingCart(); 
+    this.getPaymentTypes();
+    this.getShippingTypes();
   }
 
   checkAuthorization() {
@@ -40,7 +46,7 @@ export class CartPageComponent {
         console.log(this.shoppingCart)
       },
       (error) => {
-        this.handleError('Error fetching products', error);
+        this.errorMsg();
       }
     );
   }
@@ -56,22 +62,26 @@ export class CartPageComponent {
   deleteProductFromProductCart(productId : number, sizeTypeId : number) {
     this.cartPageService.DeleteProductFromShoppingCart(productId, this.userService.userId, sizeTypeId).subscribe(
       () => {
-        this.getShoppingCart(); //получение нового списка товаров
       },
       (error) => {
-        this.handleError('Error deleting product', error);
+        this.errorMsg();
       }
     );
   }
 
   addToOrder() {
-    this.cartPageService.AddToOrder(this.userService.userId, this.paymentTypeId, 1).subscribe(
+    if(this.paymentTypeId == 0 || this.shippingTypeId == 0){
+      this.errorMsg()
+      return;
+    }     
+
+    this.cartPageService.AddToOrder(this.userService.userId, this.paymentTypeId, this.shippingTypeId).subscribe(
       () => {
         this.deleteAllProductsFromShoppingCart();
-        this.getShoppingCart(); //получение нового списка товаров
+        this.ngOnInit(); // имба мув (движение)
       },
       (error) => {
-        this.handleError('Error deleting product', error);
+        this.errorMsg();
       }
     );
   }
@@ -81,7 +91,29 @@ export class CartPageComponent {
       () => {
       },
       (error) => {
-        this.handleError('Error deleting product', error);
+        this.errorMsg();
+      }
+    );
+  }
+
+  getPaymentTypes() {
+    this.userService.getPaymentTypes().subscribe(
+      (data: PaymentType[]) => {
+        this.paymentTypes = data;
+      },
+      (error) => {
+        this.errorMsg()
+      }
+    );
+  }
+
+  getShippingTypes() {
+    this.userService.getShippingTypes().subscribe(
+      (data: ShippingType[]) => {
+        this.shippingTypes = data;
+      },
+      (error) => {
+        this.errorMsg()
       }
     );
   }
@@ -90,7 +122,11 @@ export class CartPageComponent {
     return (this.configService.getFormattedPrice(price));
   }
 
-  handleError(message: string, error: any) {
-    console.error(message, error); // тут можно расписать полную логику при ошибке
+  errorMsg(){
+    this.statusMsg = 'Что-то пошло не так!';
+
+    setTimeout(() => {
+      this.statusMsg = ''
+    }, 2000)
   }
 }
